build[back]:Creado los modelos iniciales para las ordenes de los PDV

+++ database/migrations/2024_03_07_212910_support_services.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('description');
             $table->string('type');
-            $table->float('charge');
+            $table->decimal('charge', 9, 2);
             $table->timestamps();
         });
     }
